yosatsu/pyproject.toml
[project]
name = "yosatsu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "pdfminer-six>=20250506",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
]
